Fix todos to be deleted from local storage when project is deleted.  Fix confirmation for deleting a project

+++ src/events.js
@@ -3,7 +3,8 @@ import {showTodos} from "./project"
 import {makecheckbox} from "./form"
 
  let projects = JSON.parse(localStorage.getItem('projectlist')) || []
- let todos = JSON.parse(localStorage.getItem('todolist')) || []
+ 
+ let todos = JSON.parse(localStorage.getItem('todolist')) || [] 
 
 export {projects}
 export {todos}
@@ -13,8 +14,9 @@ export {todos}
 
 //creates a list of current project buttons 
 function showProjects(){
-    bottomleft.innerHTML = ''
-
+    bottomleft.innerHTML = '' 
+// NNED TO FIX THIS STILL!!!!!!!-----------------------------------------------------------------------------------------
+    let removelist = JSON.parse(localStorage.getItem('projectlist'))
     for(let i = 0; i<projects.length; i++){
         let currentproject = projects[i]
         let container = document.createElement('div')
@@ -25,42 +27,59 @@ function showProjects(){
         let deleteproject = document.createElement('button')
         container.append(deleteproject)
         deleteproject.setAttribute('id', 'deleteproject')
+        //delete button actions
         deleteproject.addEventListener('click',()=>{
-            console.log(projects[i].name + " has  been clicked")
-            let removelist = JSON.parse(localStorage.getItem('projectlist'))
-            let removetodos = JSON.parse(localStorage.getItem('todolist'))
-            // Need to remove the project name and update the localstorage
-            for(let i =0; i<todos.length;i++){
-                if(projects[i].name == todos[i].project){
-                    console.log(todos[i].name + "There is a match")
-                    removetodos.splice(i,1)
-                    todos.splice(i,1)
-                    localStorage.setItem('todolist', JSON.stringify(todos)) 
-                    showTodos()                
-
-
+          let answer = confirm("Are you sure?")
+          if(answer == true){
+          
+            let blanktodos = JSON.parse(localStorage.getItem('todolist'))
+            
+            // deletes the project from the localstorage
+                if(currentproject == projects[i]){
+                    projects.splice(i,1)
+                    removelist.splice(i,1)
+                    localStorage.setItem('projectlist', JSON.stringify(projects))                 
+                    showProjects()
                 }
-            }
-            for(let i =0; i<projects.length; i++){
-                if(currentproject == projects[i])
-                projects.splice(i,1)
-                removelist.splice(i,1)
-                localStorage.setItem('projectlist', JSON.stringify(projects))                 
-                showProjects()
+                //use filter metho
+
+                for (let j = blanktodos.length -1; j >= 0; j--){
+                blanktodos.splice([j],1);
+                todos.splice(j,1)
+                localStorage.setItem('todolist', JSON.stringify(todos))
+                }
+                showTodos(currentproject)
                 
+
+                    //for(let j =0; (j<blanktodos.length); j++){
+
+                  //  if(currentproject.name ==blanktodos[j].project){
+                   //     blanktodos.splice(j,1)
+                   //     todos.splice(j,1)
+                   //     localStorage.setItem('todolist', JSON.stringify(todos))
+                   //     showTodos(currentproject)
+
+                 //   }
+               // }
+
             }
-            //do same for any todos that match for
-           
+            else{return}
+        }
         
-        })
+            )
+            
+
+
+
+
+
         project.innerHTML = currentproject.name
         bottomleft.append(container)
-    
         project.addEventListener('click', ()=>{
             showTodos(currentproject.name)
             })
-        }
-}
+        
+    }}
 export{showProjects}
 
 let todobutton = document.getElementById('addtodo')
